Add wagtail-specific fixes to the CSS
Wagtail rich-text blocks come with some html that we can't get rid of,
so we need to add some CSS to fix the default <div> padding for those
richtext elements.

We combine the wagtail fixes with the nhsuk frontend library css by
adding a simple concatenation to the build step in setup.py

For now, concatenating the CSS files is ok but if any more complex fixes
are needed, we might have to start compiling the frontend library from
source.